issue #139 refactor: create reusable schedule form date handle logic to convertSchedule.js
